+++ test/support/http.js
@@ -1,20 +1,13 @@
 
-process.env.NODE_ENV = 'test';
-
 /**
  * Module dependencies.
  */
 
 var EventEmitter = require('events').EventEmitter
   , methods = ['get', 'post', 'put', 'delete', 'head']
-  , connect = require('../../')
   , http = require('http');
 
 module.exports = request;
-
-connect.proto.request = function(){
-  return request(this);
-};
 
 function request(app) {
   return new Request(app);